Implemented OBJ Loader
Created OBJLoader namespace with single loadOBJ function that takes a
filename and parses the OBJ data

Created OBJ struct and MTL struct to represent an OBJ file's
corresponding data types

MTL struct supports taking primitive data (vec3) and filenames for
textures

Updated files throughout to support uint32_t for index buffer to support
large models with more than 65536 vertices

+++ bender/bender-base/app/src/main/jni/mesh.cc
@@ -24,7 +24,7 @@ Mesh::Mesh(Renderer *renderer,
 Mesh::Mesh(Renderer *renderer,
            std::shared_ptr<Material> material,
            const std::vector<float> &vertexData,
-           const std::vector<uint16_t> &indexData) :
+           const std::vector<uint32_t> &indexData) :
     Mesh(renderer,
          material,
          std::make_shared<Geometry>(renderer->getDevice(), vertexData, indexData)) {}
